requirements.txt
aiohttp==3.6.1
appnope==0.1.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
cairocffi==1.1.0
CairoSVG==2.4.2
canvasapi==0.14.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cssselect2==0.2.2
decorator==4.4.0
defusedxml==0.6.0
dominate==2.4.0
entrypoints==0.3
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-OAuthlib==0.9.5
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
html5lib==1.0.1
httplib2==0.13.1
idna==2.8
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
mistune==0.8.4
multidict==4.5.2
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.2
oauth==1.0.1
oauth2==1.9.0.post1
oauthlib==2.1.0
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pdf-reports==0.2.4
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
-e git+https://github.com/ucfcdl/pylti.git@f98b17229c3866ea66a15eb50e79af13b980e9af#egg=PyLTI
pymongo==3.9.0
Pyphen==0.9.5
pypugjs==5.8.1
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
requests-oauthlib==1.2.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.3
SQLAlchemy==1.3.8
terminado==0.8.2
testpath==0.4.2
tinycss2==1.0.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.6
visitor==0.1.3
wcwidth==0.1.7
WeasyPrint==50
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
yarl==1.3.0
